refactor Google Calendar client
