+++ backend/app/calendar/CalendarClient.js
@@ -1,29 +1,28 @@
-const { google } = require( 'googleapis' );
-const { authorize } = require( './google-auth' );
+const Connection = require( './Connection' );
 
 class CalendarClient {
    constructor() {
       if ( CalendarClient.connection === null ) {
-         CalendarClient.connection = google.calendar( { version: 'v3', auth: authorize() } );
+         CalendarClient.connection = new Connection();
       }
 
-      this.connection = CalendarClient.connection;
+      this.calendar = CalendarClient.connection.getConnection();
    }
 
    async listCalendars() {
-      const res = await this.connection.calendarList.list( {} );
+      const res = await this.calendar.calendarList.list( {} );
 
       return res.data.items;
    }
 
    async getCalendar( calendarId ) {
-      const res = await this.connection.calendarList.get( { calendarId } );
+      const res = await this.calendar.calendarList.get( { calendarId } );
 
       return res.data;
    }
 
    async getEvents( calendarId, startDate ) {
-      const res = await this.connection.events.list( {
+      const res = await this.calendar.events.list( {
          calendarId,
          timeMin: startDate.format(),
          timeMax: startDate.add( 1, 'week' ).format(),
@@ -36,7 +35,7 @@ class CalendarClient {
    }
 
    async addEvent( event ) {
-      const overlapping = await this.connection.events.list( {
+      const overlapping = await this.calendar.events.list( {
          calendarId: event.calendar,
          timeMin: event.start.format(),
          timeMax: event.end.format(),
@@ -48,7 +47,7 @@ class CalendarClient {
          throw new Error( 'Overlapping events!' );
       }
 
-      const res = await this.connection.events.insert( {
+      const res = await this.calendar.events.insert( {
          calendarId: event.calendar,
          resource: {
             start: {
@@ -68,4 +67,4 @@ class CalendarClient {
 
 CalendarClient.connection = null;
 
-module.exports = { CalendarClient };
+module.exports = CalendarClient;
