+++ backend/bin/get-google-token.js
@@ -1,3 +1,5 @@
+#!/usr/bin/env node
+
 const readline = require( 'readline' );
 const { google } = require( 'googleapis' );
